+++ src/interfaces/CustomersSource.d.ts
@@ -1,6 +1,6 @@
 import CustomerDTO from "../core/customers/dto/CustomerDTO";
 
-export interface CustomersSystem {
+export interface CustomersSource {
   findByCPF(cpf: string): Promise<CustomerDTO | undefined>;
   findByID(id: number): Promise<CustomerDTO | undefined>;
 }
