+++ src/interfaces/ProductsSource.d.ts
@@ -1,3 +1,3 @@
-export interface ProductSystem {
+export interface ProductsSource {
   findById(id: number): Promise<ProductDTO | undefined>;
 }
